refactor: share observation timeline span primitives
